Use a probe to trigger leak detector rather than double test run
This works as follows: After running the test, the fixture now
intentionally leaks a (small) buffer, marked as a probe. It then runs
the garbage collector and polls the leak detector in a loop until it
reports a leak. If there is a legitimate pending leak from the test
run, it should always be reported before the probe since leaked
references go through a queue. Otherwise, only the probe will be
reported. Thus, if there are no leaks left after filtering out probes,
there can't have been a leak in the test run.